lua: Extract json encoder into it's own file
